#19: Move tests to concepts package and change PUT to POST
